refacto(publish): Extract publishing function
