+++ src/commands/run.js
@@ -6,15 +6,8 @@ const { NodeHtmlMarkdown } = require('node-html-markdown')
 const html2markdown = new NodeHtmlMarkdown()
 const CLI = require('clui')
 const Spinner = CLI.Spinner
-const {
-    displayError,
-    displaySuccess,
-    isDataURL,
-    validatePlatforms,
-} = require('../utils')
-const postToDev = require('./platforms/dev')
-const postToHashnode = require('./platforms/hashnode')
-const postToMedium = require('./platforms/medium')
+const { displayError, isDataURL, validatePlatforms } = require('../utils')
+const publish = require('../commands/platforms/publish.js')
 const uploadToCloudinary = require('./platforms/cloudinary')
 
 const configstore = new Conf(),
@@ -84,13 +77,19 @@ async function sourceRemotePost(
                         }
                     }
                 }
-                publish(platforms, {
-                    title,
-                    markdown,
-                    url,
-                    image,
-                    public,
-                })
+                publish(
+                    platforms,
+                    {
+                        title,
+                        markdown,
+                        url,
+                        image,
+                        public,
+                    },
+                    (message) => loading.message(message),
+                    () => loading.stop(),
+                    handleError
+                )
             } else {
                 throw new Error('No articles found in the URL.')
             }
@@ -98,54 +97,6 @@ async function sourceRemotePost(
         .catch(handleError)
 }
 
-async function publish(
-    chosenPlatforms,
-    { title, markdown, url, image, public }
-) {
-    /**
-     * Function to run after posting on a platform is done
-     */
-    function afterPost({ success, url = '', platform = '', public = false }) {
-        if (success) {
-            console.log(
-                displaySuccess(
-                    `Article ${
-                        public ? 'published' : 'added to drafts'
-                    } on ${platform} at ${url}`
-                )
-            )
-        }
-        platformsPosted++
-        checkIfShouldStopLoading(chosenPlatforms)
-    }
-
-    chosenPlatforms.forEach((platform) => {
-        switch (platform) {
-            case 'dev':
-                loading.message(`Posting article to dev.to...`)
-                postToDev(title, markdown, url, image, public, afterPost).catch(
-                    handleError
-                )
-                break
-            case 'hashnode':
-                loading.message(`Posting article to Hashnode...`)
-                postToHashnode(
-                    title,
-                    markdown,
-                    url,
-                    image,
-                    public,
-                    afterPost
-                ).catch(handleError)
-                break
-            case 'medium':
-                loading.message(`Posting article to Medium...`)
-                postToMedium(title, markdown, url, public, afterPost).catch(
-                    handleError
-                )
-        }
-    })
-}
 /**
  *
  * @param {string} url URL of the blog post
@@ -188,15 +139,6 @@ async function run(
 function handleError(err) {
     loading.stop()
     console.error(displayError(err))
-}
-
-/**
- * If the number of platforms posted on is complete stop the loading
- */
-async function checkIfShouldStopLoading(chosenPlatforms) {
-    if (platformsPosted === chosenPlatforms.length) {
-        loading.stop()
-    }
 }
 
 /**
